pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Demonstrate LLM instrumentation"
authors = [
    { name = "Tim Van Wassenhove", email = "github@timvw.be" }
]
dependencies = [
    "openai>=1.51.2",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "arize-phoenix>=5.2.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/demo"]
